Address review feedback: Update documentation for component version management, split resolve into subcommands and implement hash in merging logic.
