+++ pkg/apis/config/v1alpha1/validation/validation_test.go
@@ -286,7 +286,7 @@ var _ = Describe("Validation", func() {
 			It("should pass with a valid DefaultVersionsUpdateStrategy", func() {
 				conf := &v1alpha1.LandscapeKitConfiguration{
 					VersionConfig: &v1alpha1.VersionConfiguration{
-						DefaultVersionsUpdateStrategy: new(v1alpha1.DefaultVersionsUpdateStrategy("ReleaseBranch")),
+						DefaultVersionsUpdateStrategy: new(v1alpha1.DefaultVersionsUpdateStrategyReleaseBranch),
 					},
 				}
