+++ pkg/utils/meta/diff.go
@@ -6,10 +6,10 @@ package meta
 
 import (
 	"bytes"
+	"crypto/sha256"
 	"fmt"
 	"maps"
 	"slices"
-	"strings"
 
 	"github.com/elliotchance/orderedmap/v3"
 	"go.yaml.in/yaml/v4"
@@ -138,8 +138,12 @@ func buildKey(t map[string]any) string {
 	namespace, _ := metadata["namespace"].(string)
 	if typeKey == "" && namespace == "" && name == "" {
 		keys := slices.Collect(maps.Keys(t))
+		if len(keys) == 0 {
+			return ""
+		}
 		slices.Sort(keys)
-		return strings.Join(keys, "-")
+		sum := sha256.Sum256(fmt.Appendf(nil, "%v", keys))
+		return fmt.Sprintf("%x", sum)
 	}
 
 	return typeKey + "/" + namespace + "/" + name
